Handle the volumes created by the CSI driver (#109)
Current VolumeSnapshotter works with volumes allocated by the "in-tree" storage driver. Kubernetes is moving to the CSI driver for Azure and will deprecate the in-tree driver, however we may not be able to get CSI snapshots in Velero to GA before that happens. As an interim solution, we handle volumes created by the CSI driver with the existing VolumeSnapshotter.

Fixes https://github.com/vmware-tanzu/velero/issues/4109

Signed-off-by: Wenkai Yin(尹文开) <yinw@vmware.com>